upgradepkg: de-duplicate code calling removepkg
ROOT variable is already exported so there is no need to explicitly pass
it to removepkg.  Furthermore, by unsetting ROOT if it's not a directory,
we can simply use "$ROOT/var/log/packages" to mean either directory under
$ROOT or one under /.  And as a matter of fact, there is no need to change
directory since all we need is a list of files in the directory.  Lastly,
double-grep pipe can be replaced by a simple regex.